+++ tests/testthat/test-zero-networks.R
@@ -13,9 +13,9 @@ zero_units_rows <- c(1,3,5);
 zero_units <- accum$connection.counts[zero_units_rows,]$ENA_UNIT
 accum$connection.counts[ENA_UNIT %in% zero_units, 8:22] = 0
 
-test_that("no force zero networks", {
+test_that("testing zero networks", {
   ### create ENA set WITHOUT center alignment
-  set_F = ena.make.set(enadata = accum);
+  set_F = ena.make.set(enadata = accum, center.align.to.origin = FALSE);
 
   # sapply(as.matrix(set_F$points[ENA_UNIT %in% zero_units,]), all.equal)
 
@@ -25,9 +25,7 @@ test_that("no force zero networks", {
 
   mean_centroid_F = colMeans(set_F$model$centroids[,2:ncol(set_F$model$centroids)])
   testthat::expect_true(all(mean_centroid_F != 0))
-})
 
-test_that("force zero points", {
   set_T = ena.make.set(
     enadata = accum,
     center.align.to.origin = TRUE
@@ -42,4 +40,10 @@ test_that("force zero points", {
   ### And the mean centroid is also zero
   mean_centroid_T = colMeans(as.matrix(set_T$model$centroids))
   testthat::expect_equal(sum(mean_centroid_T), 0)
+
+  centroid_compare <- set_T$model$centroids == set_F$model$centroids
+  testthat::expect_true(all(centroid_compare[zero_units_rows,]))
+  testthat::expect_true(all(centroid_compare[,1]))
+  testthat::expect_true(all(centroid_compare[-zero_units_rows,-1] == FALSE))
+  testthat::expect_true(all((set_T$rotation$nodes == set_F$rotation$nodes)[,-1] == FALSE))
 })